requirements.txt
slackclient>=2.7.1
slackeventsapi>=2.1.0